v1.0 it's really working

+++ src/main/kotlin/dto/ResultDto.kt
@@ -1,12 +1,10 @@
 package dto
 
-import java.time.LocalDateTime
-
 data class ResultDto(
     val x: Double,
     val y: Double,
     val r: Double,
     val hit: Boolean,
-    val execTimeMs: Long,
+    val execTimeMs: Double,
     val createdAt: String
 )
